Move to top-level directory
Was able to make UI mobile-friendly with inline conditions rather than
separate components, so removing this directory organization.
