+++ ui/settings.js
